Chore: refactoring & unit tests in progress
