+++ bc_obps/common/service/email/email_service.py
@@ -1,9 +1,9 @@
 from datetime import datetime, timedelta
 from typing import List, Optional
 from uuid import UUID
-from common.enums import AccessRequestStates, OperatorAccessRequest, AccessRequestTypes
 import logging
 import requests
+from registration.enums.enums import AccessRequestStates, OperatorAccessRequest, AccessRequestTypes
 from common.models import EmailNotification, EmailNotificationTemplate
 from django.conf import settings
 
@@ -218,51 +218,3 @@ class EmailService(object):
             recipients_email=recipients_email,
             template_id=template_id,
         )
-
-    def send_operator_access_request_email(
-        self,
-        access_state: AccessRequestStates,
-        access_type: AccessRequestTypes,
-        operator_legal_name: str,
-        external_user_full_name: str,
-        external_user_email_address: str,
-    ) -> None:
-        """
-        Sends an email to a user regarding an operator and their access request based on the access state and type.
-
-        Args:
-            access_state: The state of the operator access request.
-            access_type: The type of the operator for the access request.
-            operator_legal_name: The legal name of the operator to use in the email template.
-            external_user_full_name: The full name of the external user to use in the email template.
-            external_user_email_address: The email address of the external user to use in the email template.
-
-        Raises:
-            ValueError: If the email template is not found.
-
-        Returns:
-            None
-        """
-
-        try:
-            template_name = OperatorAccessRequest.generate_access_request_template_name(
-                access_type.value, access_state.value
-            )
-            template = EmailNotificationTemplate.objects.get(name=template_name)
-        except EmailNotificationTemplate.DoesNotExist:
-            raise ValueError("Email template not found")
-
-        email_context = {
-            "operator_legal_name": operator_legal_name,
-            "external_user_full_name": external_user_full_name,
-            "external_user_email_address": external_user_email_address,
-        }
-
-        try:
-            response_json = self.send_email_by_template(template, email_context, [external_user_email_address])
-            # Create email notification record to store the transaction and message IDs
-            self.create_email_notification_record(
-                response_json['txId'], response_json['messages'][0]['msgId'], [external_user_email_address], template.id
-            )
-        except Exception as exc:  # not raising exception here to avoid breaking the flow of the application
-            logger.error(f'Logger: Exception sending operator access request email - {str(exc)}')
