+++ bc_obps/registration/tests/endpoints/select_operator/test_request_admin_access.py
@@ -1,4 +1,4 @@
-from common.enums import AccessRequestStates, AccessRequestTypes
+from registration.enums.enums import AccessRequestStates, AccessRequestTypes
 from common.service.email.email_service import EmailService
 from registration.models import (
     UserOperator,
